+++ configs/db/tables/gems__radius_config.999.sql
@@ -5,7 +5,10 @@ CREATE TABLE if not exists gems__radius_config (
         grcfg_ip                varchar(39) CHARACTER SET 'utf8' collate utf8_unicode_ci default NULL,
         grcfg_port              int(5) default NULL,
         grcfg_secret            varchar(255) CHARACTER SET 'utf8' collate utf8_unicode_ci default NULL,
+
+        -- deprecated in 1.8.6  method was never used, now saved with password
         grcfg_encryption        varchar(20) CHARACTER SET 'utf8' COLLATE 'utf8_general_ci' null,
+        -- end deprecated
 
         PRIMARY KEY (grcfg_id)
     )